Move common code up into BaseGridView; remove identical subclasses
This also follows the Law of Demeter a litle better.
