+++ src/main/java/com/refinedmods/refinedstorage/screen/grid/stack/FluidGridStack.java
@@ -155,7 +155,16 @@ public class FluidGridStack implements IGridStack {
     @Override
     public int getQuantity() {
         // The isCraftable check is needed so sorting is applied correctly
-        return isCraftable() ? 0 : stack.getAmount();
+        return isCraftable() || zeroed ? 0 : stack.getAmount();
+    }
+
+    @Override
+    public void setQuantity(int amount) {
+        if (amount <= 0) {
+            setZeroed(true);
+        } else {
+            stack.setAmount(amount);
+        }
     }
 
     @Override
