+++ src/main/java/com/refinedmods/refinedstorage/screen/grid/stack/IGridStack.java
@@ -30,6 +30,8 @@ public interface IGridStack {
 
     int getQuantity();
 
+    void setQuantity(int amount);
+
     String getFormattedFullQuantity();
 
     void draw(MatrixStack matrixStack, BaseScreen<?> screen, int x, int y);
