+++ src/main/java/com/refinedmods/refinedstorage/screen/grid/view/GridViewImpl.java
@@ -13,7 +13,7 @@ import java.util.function.Predicate;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-public abstract class BaseGridView implements IGridView {
+public class GridViewImpl implements IGridView {
     private final GridScreen screen;
     private boolean canCraft;
 
@@ -23,7 +23,7 @@ public abstract class BaseGridView implements IGridView {
     private List<IGridStack> stacks = new ArrayList<>();
     protected final Map<UUID, IGridStack> map = new HashMap<>();
 
-    protected BaseGridView(GridScreen screen, IGridSorter defaultSorter, List<IGridSorter> sorters) {
+    public GridViewImpl(GridScreen screen, IGridSorter defaultSorter, List<IGridSorter> sorters) {
         this.screen = screen;
         this.defaultSorter = defaultSorter;
         this.sorters = sorters;
@@ -97,6 +97,54 @@ public abstract class BaseGridView implements IGridView {
             };
         } else {
             return filters;
+        }
+    }
+
+    @Override
+    public void setStacks(List<IGridStack> stacks) {
+        map.clear();
+
+        for (IGridStack stack : stacks) {
+            map.put(stack.getId(), stack);
+        }
+    }
+
+    @Override
+    public void postChange(IGridStack stack, int delta) {
+        // COMMENT 1 (about this if check in general)
+        // Update the other id reference if needed.
+        // Taking a stack out - and then re-inserting it - gives the new stack a new ID
+        // With that new id, the reference for the crafting stack would be outdated.
+
+        // COMMENT 2 (about map.containsKey(stack.getOtherId()))
+        // This check is needed or the .updateOtherId() call will crash with a NPE in high-update environments.
+        // This is because we might have scenarios where we process "old" delta packets from another session when we haven't received any initial update packet from the new session.
+        // (This is because of the executeLater system)
+        // This causes the .updateOtherId() to fail with a NPE because the map is still empty or the IDs mismatch.
+        // We could use !map.isEmpty() here too. But if we have 2 "old" delta packets, it would rightfully ignore the first one. But this method mutates the map and would put an entry.
+        // This means that on the second delta packet it would still crash because the map wouldn't be empty anymore.
+        if (!stack.isCraftable() &&
+                stack.getOtherId() != null &&
+                map.containsKey(stack.getOtherId())) {
+            IGridStack craftingStack = map.get(stack.getOtherId());
+
+            craftingStack.updateOtherId(stack.getId());
+            craftingStack.setTrackerEntry(stack.getTrackerEntry());
+        }
+
+        IGridStack existing = map.get(stack.getId());
+
+        if (existing == null) {
+            stack.setQuantity(delta);
+
+            map.put(stack.getId(), stack);
+        } else {
+            existing.setQuantity(existing.getQuantity() + delta);
+            if (existing.getQuantity() <= 0) {
+                map.remove(existing.getId());
+            }
+
+            existing.setTrackerEntry(stack.getTrackerEntry());
         }
     }
